+++ CHANGELOG.md
@@ -13,3 +13,7 @@
 7.1.4: Save new slug map items in the "post flush" event instead of "post persist".
 
 7.1.5: Filterer: allow to filter by "many-to-many" associations.
+
+7.2.0: Dispatch filterer build constraint event.
+
+7.2.1: Allow to blacklist object classes in "Darvin\ContentBundle\Repository\SlugMapItemRepository::getBySlugsChildren()".